Move facility title and view route button to separate components
